pyproject.toml
[tool.poetry]
name = "orchestrator-kafka-proxy"
version = "0.1.0"
description = ""
authors = ["Michele Perniola <michele.perniola@ba.infn.it>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
kafka-python = "2.0.3"
python-dateutil = "^2.8.2"
pytz = "2025.1"
python-json-config = "1.2.3"
Flask = "3.0.2"
Werkzeug = "3.0.3"
ruff = "0.11.5"